Change pegout ID description (#375)
* Use new pegout id definition
Pegout ID should be the hash of:

Version number
Sequential number (incremental)
stream id
packet number
slot id que (0-99)
operator take pub key
random number (bitcoin latest block hash)

* Fix tests after pegoutId change

* Fix slither reentrancy warning

* Fix PR comments

* Renamed getOperatorDisputeData to getOperatorData

* Add Pegout ID documentation

* Get back to use for loop in getOperatorData

* Add operatorTakePubKey to TempPegoutInfo and event AdvanceFundsRegistered

* Update events for zk proof

* Update getOperatorDisputeData to selectTakeOperator
This way it's clearer that it selects and change state besides returning the selecter member pub keys

* Clean up

+++ script/AddOperatorTakeTxid.s.sol
@@ -67,7 +67,7 @@ contract addOperatorTakeTxidsScript is ScriptUtils, ContractAddressManager {
         setUp(_mnemonicIndex, _acceptPeginTxid);
 
         // REIMBURSEMENT KICKOFF
-        BtcTransaction memory kickoffTx = createReimbursementKickoffTx(committeePubKey, uint32(expectedSlotId));
+        BtcTransaction memory kickoffTx = createReimbursementKickoffTx(committeePubKey, expectedSlotId);
         bytes32 reimbursementKickoffTxid = getTxid(kickoffTx);
 
         // OPERATOR TAKE
