+++ script/helpers/ScriptUtils.sol
@@ -307,7 +307,7 @@ abstract contract ScriptUtils is Script {
         return BtcTransaction({version: Constants.BTC_TX_VERSION, inputs: btcInputs, outputs: btcOutputs, locktime: 0});
     }
 
-    function createReimbursementKickoffTx(bytes memory _committeePubKey, uint32 _slotIndex)
+    function createReimbursementKickoffTx(bytes memory _committeePubKey, uint64 _slotIndex)
         internal
         pure
         returns (BtcTransaction memory)
@@ -317,7 +317,7 @@ abstract contract ScriptUtils is Script {
         btcInputs[0] = BtcTxIn({
             // Input txid is uncheckable by the contract
             txId: bytes32(0xabcdefabcdefabcdefabcdefabcdefabcdefabcdefabcdefabcdefabcdefabcd),
-            vout: _slotIndex,
+            vout: uint32(_slotIndex),
             sequence: Constants.SEQUENCE,
             scriptSig: hex""
         });
